+++ system/lib/compiler-rt/lib/asan/asan_shadow_setup.cc
@@ -14,9 +14,9 @@
 
 #include "sanitizer_common/sanitizer_platform.h"
 
-// asan_fuchsia.cc and asan_rtems.cc have their own
+// asan_fuchsia.cc, asan_rtems.cc and asan_emscripten.cc have their own
 // InitializeShadowMemory implementation.
-#if !SANITIZER_FUCHSIA && !SANITIZER_RTEMS
+#if !SANITIZER_FUCHSIA && !SANITIZER_RTEMS && !SANITIZER_EMSCRIPTEN
 
 #include "asan_internal.h"
 #include "asan_mapping.h"